--- src/controls/LevelingSensor/QsgNodeUtility.cpp
@@ -1,1 +1,0 @@
-#include "QsgNodeUtility.h"
